requirements.txt
Flask>=3.0.0
Pillow>=10.3.0
numpy>=1.26.0
Werkzeug>=3.0.0
